Refactor - Split Selector methods
