+++ src/DynamicQueryable/DynamicQueryable.cs
@@ -10,96 +10,6 @@ namespace System.Linq.Dynamic {
             return (IQueryable<T>)source;
         }
 
-        public static IQueryable Select(this IQueryable source, string selector, params object[] values) {
-            return Select(source, selector, null, values);
-        }
-
-        public static IQueryable Select(this IQueryable source, string selector, IDictionary<string, object> variables, params object[] values) {
-            if (source == null) throw new ArgumentNullException(nameof(source));
-            if (selector == null) throw new ArgumentNullException(nameof(selector));
-
-            var lambda = Evaluator.ToLambda(selector, new[] { source.ElementType }, variables, values);
-
-            return source.Provider.CreateQuery(
-                Expression.Call(
-                    typeof(Queryable),
-                    "Select",
-                    new[] { source.ElementType, lambda.Body.Type },
-                    source.Expression,
-                    Expression.Quote(lambda)
-                )
-            );
-        }
-
-        public static IQueryable<T> SelectMany<T>(this IQueryable source, string selector, params object[] values) {
-            return SelectMany<T>(source, selector, null, values);
-        }
-
-        public static IQueryable<T> SelectMany<T>(this IQueryable source, string selector, Dictionary<string, object> variables, params object[] values) {
-            return (IQueryable<T>)SelectMany((IQueryable)source, selector, null, values);
-        }
-
-        public static IQueryable SelectMany(this IQueryable source, string selector, params object[] values) {
-            return SelectMany(source, selector, null, values);
-        }
-
-        public static IQueryable SelectMany(this IQueryable source, string selector, Dictionary<string, object> variables, params object[] values) {
-            if (source == null) throw new ArgumentNullException(nameof(source));
-            if (selector == null) throw new ArgumentNullException(nameof(selector));
-
-            var lambda = Evaluator.ToLambda(selector, new[] { source.ElementType }, variables, values);
-
-            // Fix lambda by recreating to be of correct Func<> type in case 
-            // the expression parsed to something other than IEnumerable<T>.
-            // For instance, a expression evaluating to List<T> would result 
-            // in a lambda of type Func<T, List<T>> when we need one of type
-            // an Func<T, IEnumerable<T> in order to call SelectMany().
-            var inputType = source.Expression.Type.GetGenericArguments()[0];
-            var resultType = lambda.Body.Type.GetGenericArguments()[0];
-            var enumerableType = typeof(IEnumerable<>).MakeGenericType(resultType);
-            var delegateType = typeof(Func<,>).MakeGenericType(inputType, enumerableType);
-            lambda = Expression.Lambda(delegateType, lambda.Body, lambda.Parameters);
-
-            return source.Provider.CreateQuery(
-                Expression.Call(
-                    typeof(Queryable),
-                    "SelectMany",
-                    new[] { source.ElementType, resultType },
-                    source.Expression,
-                    Expression.Quote(lambda)
-                )
-            );
-        }
-
-        public static IQueryable<T> OrderBy<T>(this IQueryable<T> source, string ordering, params object[] values) {
-            return OrderBy(source, ordering, null, values);
-        }
-
-        public static IQueryable<T> OrderBy<T>(this IQueryable<T> source, string ordering, Dictionary<string, object> variables, params object[] values) {
-            return (IQueryable<T>)OrderBy((IQueryable)source, ordering, variables, values);
-        }
-
-        public static IQueryable OrderBy(this IQueryable source, string ordering, params object[] values) {
-            return OrderBy(source, ordering, null, values);
-        }
-
-        public static IQueryable OrderBy(this IQueryable source, string ordering, Dictionary<string, object> variables, params object[] values) {
-            if (source == null) throw new ArgumentNullException(nameof(source));
-            if (ordering == null) throw new ArgumentNullException(nameof(ordering));
-
-            var lambda = Evaluator.ToLambda(ordering, new[] { source.ElementType }, variables, values);
-
-            return source.Provider.CreateQuery(
-                Expression.Call(
-                    typeof(Queryable),
-                    "OrderBy",
-                    new[] { source.ElementType, lambda.ReturnType },
-                    source.Expression,
-                    Expression.Quote(lambda)
-                )
-            );
-        }
-
         public static IQueryable Take(this IQueryable source, int count) {
             if (source == null) throw new ArgumentNullException(nameof(source));
